Implement if statements
