+++ example/first_exe.sl
@@ -1,4 +1,5 @@
-(extern int print_double (d: float))  ;; declare external print_double function
+;; declare external print_double function
+(extern int print_double (d: float))
 
 (defun foo (x: float, y: float)
   (* (+ x y) 4.0))
